Move mod_info_panel.py into RimSort.views.sub_views
