+++ RimSort/views/main_content_panel.py
@@ -63,7 +63,7 @@ from sort.topo_sort import *
 from RimSort.views.sub_views.actions_panel import Actions
 from sub_view.active_mods_panel import ActiveModList
 from sub_view.inactive_mods_panel import InactiveModList
-from sub_view.mod_info_panel import ModInfo
+from RimSort.views.sub_views.mod_info_panel import ModInfo
 from util.metadata import *
 from util.schema import validate_mods_config_format
 from util.steam.steamcmd.wrapper import SteamcmdInterface
