+++ RimSort/views/sub_views/mod_info_panel.py
